Move signAndSendTransactionsInSeries into utils folder
