+++ src/utils/transactions/helpers/index.ts
@@ -2,3 +2,4 @@ export * from './signAndSendTransaction';
 export * from './showSolscanLinkNotification';
 export * from './getArrayByNumber';
 export * from './signAndSendAllTransactions';
+export * from './signAndSendTransactionsInSeries';
